PyTorch/transformer/attention_is_all_you_need/requirements.txt
torchtext==0.14.1
dill==0.3.4
tqdm
iopath
numpy
